+++ contracts/solana-multisig-prover/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod contract;
+pub mod error;
+pub mod events;
+mod flags;
+mod state;
+
+#[cfg(test)]
+mod test;