+++ integration-tests/src/lib.rs
@@ -8,6 +8,7 @@ pub mod gateway_contract;
 pub mod multisig_contract;
 pub mod multisig_prover_contract;
 pub mod protocol;
+pub mod solana_multisig_prover_contract;
 pub mod rewards_contract;
 pub mod router_contract;
 pub mod service_registry_contract;